Modified display of units so that "Base Unit" is the the unit with index 0. Also added visual for conversion formula.
