+++ Client/src/Components/Profile/UnitManagementSection/UnitForm.tsx
@@ -40,12 +40,6 @@ export const UnitForm = (props: IProps) => {
     <>
       <Box className={classStyles().defaultBorder}>
         <ThemeProvider theme={disabledTheme}>
-          <Typography variant='h6' align="left">Base Unit ID:</Typography>
-          <Grid container spacing={4}>
-            <Grid item sm={4}>
-              <FastField name="baseUnitId" label='Base Unit ID' component={MuiTextFieldFormik} />
-            </Grid>
-          </Grid>
           <Grid container spacing={4}>
             <FieldArray name='units' >
               {({ form, ...fieldArrayHelpers }) => <UnitList units={form.values.units} fieldArrayHelpers={fieldArrayHelpers} />}
@@ -53,8 +47,6 @@ export const UnitForm = (props: IProps) => {
           </Grid>
         </ThemeProvider>
       </Box >
-      <Divider className={classStyles().divider} variant="middle" />
-      <Typography align="left">Example Conversion:</Typography>
     </>
   )
